tidy up KNOW_LINKS and KNOWN_LINKS_PATTERNS to constants/tools.jsx
